refactor: 선택된 쿠폰 지역상태로 관리하는 것으로 변경
