+++ src/components/OrderInfo/index.ts
@@ -1,4 +1,4 @@
-export { default as Coupon } from './Coupon/Coupon';
+export { default as CouponAndPaymentInfo } from './Coupon/CouponAndPaymentInfo';
 export { default as IsolatedRegionShippingFee } from './IsolatedRegionShippingFee/IsolatedRegionShippingFee';
 export { default as PaymentTotalWithDiscount } from './PaymentTotalWithDiscount/PaymentTotalWithDiscount';
 export { default as ShoppingCartItemView } from './ShoppingCartItemView/ShoppingCartItemView';
